requirements.txt
annotated-types==0.7.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.9.2
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==6.0.0
certifi==2025.4.26
cffi==2.0.0
channels==4.3.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
contourpy==1.3.2
crispy-bootstrap5==2025.6
cryptography==46.0.3
cupshelpers==1.0
cycler==0.12.1
daphne==4.2.1
dbus-python==1.2.18
defer==1.0.6
dill==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==3.0.1
Django==5.1.7
django-crispy-forms==2.4
django-extensions==4.1
django-oauth-toolkit==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
dnslib==0.9.26
dnspython==2.7.0
drf-yasg==1.21.10
duplicity==0.8.21
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.104.1
fasteners==0.14.1
fonttools==4.57.0
future==0.18.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==24.7.2
inflection==0.5.1
jeepney==0.7.1
jiter==0.8.2
joblib==1.4.2
jwcrypto==1.5.6
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
localstack==3.2.0
localstack-core==3.2.0
localstack-ext==3.2.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.10.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.5
oauthlib==3.3.1
olefile==0.46
openai==1.63.0
packaging==25.0
pandas==2.2.3
paramiko==2.9.3
pbr==6.1.1
pexpect==4.8.0
Pillow==9.0.1
plux==1.12.1
protobuf==3.12.4
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.11
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.5.0
pydantic_core==2.14.1
pydot==4.0.1
pygame==2.6.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.9.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-jose==3.5.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
semver==3.0.4
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SpeechRecognition==3.14.2
sqlparse==0.5.3
srt==3.5.3
starlette==0.27.0
stevedore==5.4.1
systemd-python==234
tabulate==0.9.0
tailer==0.4.1
threadpoolctl==3.6.0
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.4.0
usb-creator==0.3.7
uvicorn==0.24.0
vboxapi==1.0
vosk==0.3.45
wadllib==1.3.6
websockets==15.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==8.0.1
